+++ R/imports.R
@@ -1,9 +1,9 @@
 #' @useDynLib sparsegl, .registration = TRUE
 #' @importFrom utils packageDescription
 #' @importFrom methods cbind2 rbind2 as
-#' @importFrom stats approx coef predict
-#' @importFrom rlang .data
-#' @importFrom rlang :=
+#' @importFrom stats approx coef predict fitted
+#' @importFrom rlang .data abort warn
+#' @importFrom rlang := %||%
 #' @import dotCall64 Matrix
 #' @docType package
 NULL